Fix P0 critical issues: timer polling race condition and gap_epoch overflow
1. Timer Polling Race Condition: Always use fresh rdtsc() for timer
   polling instead of stale last_poll_tsc_ to avoid missing timer
   deadlines when events arrive with timestamps out of order.

2. Upgrade gap_epoch to uint16_t: Changed from uint8_t to uint16_t in
   SequenceTracker and OrderState to avoid wrap-around issues in
   high-throughput systems with frequent gaps (wraps at 65535 instead
   of 255).

+++ src/core/order_state.hpp
@@ -66,7 +66,7 @@ struct OrderState {
     MismatchMask current_mismatch{};      // 1-byte mask
 
     std::uint32_t timer_generation{0};    // generation-based lazy cancel
-    std::uint8_t gap_suppression_epoch{0};
+    std::uint16_t gap_suppression_epoch{0};  // Upgraded to uint16_t to avoid wrap-around issues
 
     // ===== Divergence emission tracking (FX-7053) =====
     // These fields support idempotent divergence emission to avoid flooding
